refactor(server): split maintenance_fixups.go into 8 domain files
Splits the 6400-line maintenance_fixups.go into:
- maintenance_readby.go (284 lines)
- maintenance_series.go (332 lines)
- maintenance_files.go (1120 lines)
- maintenance_author_version.go (469 lines)
- maintenance_dedup.go (866 lines)
- maintenance_wipe.go (380 lines)
- maintenance_itunes.go (2565 lines)
- maintenance_hashes.go (505 lines)

No logic changes. Structure audit STRUCT-3.
